refactor: move parser and router logic to separate filed from index.js
