--- src/file/a.txt
@@ -1,1 +1,0 @@
-wow!wow!wow!
