--- src/file/fileToRead.txt
@@ -1,1 +1,0 @@
-This is file to read!
