--- src/file/fileToWrite.txt
@@ -1,1 +1,0 @@
-This is file to write
